+++ test/resources/language/terraform/error.tf
@@ -1,0 +1,4 @@
+resource "aws_instance" "web" {
+  ami           = "ami-b73b63a0"
+  instance_type = "t1.2xlarge" # invalid type
+}