+++ CHANGES.rst
@@ -32,6 +32,8 @@
   - Text prose with ``textlint`` [GH-1534]
   - VHDL with ``ghdl`` [GH-1160]
   - mypy with ``python-mypy`` [GH-1354]
+  - terraform with ``terraform fmt`` [GH-1586]
+  - terraform-tflint with ``tflint`` [GH-1586]
 
 - New features: